# BENCH, BUG: fix Savez suite, previously was actually calling pickle.dump()